+++ src/Amadeus/Client/Session/Handler/SoapHeader4.php
@@ -344,21 +344,17 @@ class SoapHeader4 extends Base
      */
     protected function generateGuid()
     {
-        if (function_exists('com_create_guid')) {
-            return com_create_guid();
-        } else {
-            mt_srand((double)microtime()*10000);
-            $charId = strtoupper(md5(uniqid(rand(), true)));
-            $hyphen = chr(45); // "-"
+        mt_srand((double)microtime()*10000);
+        $charId = strtoupper(md5(uniqid(rand(), true)));
+        $hyphen = chr(45); // "-"
 
-            $uuid = substr($charId, 0, 8) . $hyphen
-                .substr($charId, 8, 4) . $hyphen
-                .substr($charId, 12, 4) . $hyphen
-                .substr($charId, 16, 4) . $hyphen
-                .substr($charId, 20, 12);
+        $uuid = substr($charId, 0, 8) . $hyphen
+            .substr($charId, 8, 4) . $hyphen
+            .substr($charId, 12, 4) . $hyphen
+            .substr($charId, 16, 4) . $hyphen
+            .substr($charId, 20, 12);
 
-            return $uuid;
-        }
+        return $uuid;
     }
 
     /**