+++ src/Amadeus/Client/Struct/Air/MultiAvailability/ProductAvailability.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Amadeus
+ *
+ * Copyright 2015 Amadeus Benelux NV
+ */
+
+namespace Amadeus\Client\Struct\Air\MultiAvailability;
+
+/**
+ * ProductAvailability
+ *
+ * @package Amadeus\Client\Struct\Air\MultiAvailability
+ * @author Dieter Devlieghere <dieter.devlieghere@benelux.amadeus.com>
+ */
+class ProductAvailability
+{
+    /**
+     * @var string
+     */
+    public $discountCode;
+
+    /**
+     * @var int
+     */
+    public $numberOfSeats;
+}